+++ src/sora.h
@@ -13,6 +13,7 @@
 #include <boost/asio/io_context.hpp>
 
 // WebRTC
+#include <api/environment/environment_factory.h>
 #include <api/scoped_refptr.h>
 #include <api/task_queue/task_queue_factory.h>
 #include <media/engine/webrtc_media_engine.h>
@@ -80,9 +81,6 @@ class Sora : public std::enable_shared_from_this<Sora>,
   void* GetAndroidApplicationContext(void* env);
   static sora_conf::ErrorCode ToErrorCode(sora::SoraSignalingErrorCode ec);
 
-  static sora::SoraSignalingConfig::ForwardingFilter ConvertToForwardingFilter(
-      const sora_conf::internal::ForwardingFilter& filter);
-
   // SoraSignalingObserver の実装
   void OnSetOffer(std::string offer) override;
   void OnDisconnect(sora::SoraSignalingErrorCode ec,
@@ -90,10 +88,10 @@ class Sora : public std::enable_shared_from_this<Sora>,
   void OnNotify(std::string text) override;
   void OnPush(std::string text) override;
   void OnMessage(std::string label, std::string data) override;
-  void OnTrack(
-      rtc::scoped_refptr<webrtc::RtpTransceiverInterface> transceiver) override;
+  void OnTrack(webrtc::scoped_refptr<webrtc::RtpTransceiverInterface>
+                   transceiver) override;
   void OnRemoveTrack(
-      rtc::scoped_refptr<webrtc::RtpReceiverInterface> receiver) override;
+      webrtc::scoped_refptr<webrtc::RtpReceiverInterface> receiver) override;
   void OnDataChannel(std::string label) override;
 
  private:
@@ -101,24 +99,24 @@ class Sora : public std::enable_shared_from_this<Sora>,
                  std::function<void(int, std::string)> on_disconnect);
   void DoSwitchCamera(
       const sora_conf::internal::CameraConfig& cc,
-      rtc::scoped_refptr<webrtc::VideoTrackInterface> video_track);
+      webrtc::scoped_refptr<webrtc::VideoTrackInterface> video_track);
 
-  static rtc::scoped_refptr<UnityAudioDevice> CreateADM(
-      webrtc::TaskQueueFactory* task_queue_factory,
+  static webrtc::scoped_refptr<UnityAudioDevice> CreateADM(
+      webrtc::Environment env,
       bool dummy_audio,
       bool unity_audio_input,
       bool unity_audio_output,
       std::function<void(const int16_t*, int, int)> on_handle_audio,
       std::string audio_recording_device,
       std::string audio_playout_device,
-      rtc::Thread* worker_thread,
+      webrtc::Thread* worker_thread,
       void* jni_env,
       void* android_context);
-  static bool InitADM(rtc::scoped_refptr<webrtc::AudioDeviceModule> adm,
+  static bool InitADM(webrtc::scoped_refptr<webrtc::AudioDeviceModule> adm,
                       std::string audio_recording_device,
                       std::string audio_playout_device);
 
-  static rtc::scoped_refptr<webrtc::VideoTrackSourceInterface>
+  static webrtc::scoped_refptr<webrtc::VideoTrackSourceInterface>
   CreateVideoCapturer(
       int capturer_type,
       void* unity_camera_texture,
@@ -128,21 +126,22 @@ class Sora : public std::enable_shared_from_this<Sora>,
       int video_height,
       int video_fps,
       std::function<void(const webrtc::VideoFrame& frame)> on_frame,
-      rtc::Thread* signaling_thread,
+      webrtc::Thread* signaling_thread,
       void* jni_env,
       void* android_context,
       UnityContext* unity_context);
 
   void PushEvent(std::function<void()> f);
 
-  struct CapturerSink : rtc::VideoSinkInterface<webrtc::VideoFrame> {
-    CapturerSink(rtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer,
-                 std::function<void(std::string)> on_frame);
+  struct CapturerSink : webrtc::VideoSinkInterface<webrtc::VideoFrame> {
+    CapturerSink(
+        webrtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer,
+        std::function<void(std::string)> on_frame);
     ~CapturerSink() override;
     void OnFrame(const webrtc::VideoFrame& frame) override;
 
    private:
-    rtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer_;
+    webrtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer_;
     std::function<void(std::string)> on_frame_;
   };
 
@@ -151,9 +150,9 @@ class Sora : public std::enable_shared_from_this<Sora>,
   std::shared_ptr<sora::SoraSignaling> signaling_;
   UnityContext* unity_context_;
   std::unique_ptr<UnityRenderer> renderer_;
-  rtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track_;
-  rtc::scoped_refptr<webrtc::VideoTrackInterface> video_track_;
-  rtc::scoped_refptr<webrtc::RtpSenderInterface> video_sender_;
+  webrtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track_;
+  webrtc::scoped_refptr<webrtc::VideoTrackInterface> video_track_;
+  webrtc::scoped_refptr<webrtc::RtpSenderInterface> video_sender_;
   std::function<void(ptrid_t, std::string)> on_add_track_;
   std::function<void(ptrid_t, std::string)> on_remove_track_;
   std::function<void(std::string)> on_set_offer_;
@@ -166,7 +165,7 @@ class Sora : public std::enable_shared_from_this<Sora>,
   std::function<void(std::string)> on_capturer_frame_;
 
   std::shared_ptr<sora::SoraClientContext> sora_context_;
-  std::unique_ptr<rtc::Thread> io_thread_;
+  std::unique_ptr<webrtc::Thread> io_thread_;
 
   std::mutex event_mutex_;
   std::deque<std::function<void()>> event_queue_;
@@ -177,11 +176,11 @@ class Sora : public std::enable_shared_from_this<Sora>,
 
   std::string stream_id_;
 
-  rtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer_;
+  webrtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer_;
   int capturer_type_ = 0;
   std::shared_ptr<CapturerSink> capturer_sink_;
 
-  rtc::scoped_refptr<UnityAudioDevice> unity_adm_;
+  webrtc::scoped_refptr<UnityAudioDevice> unity_adm_;
   webrtc::TaskQueueFactory* task_queue_factory_;
 #if defined(SORA_UNITY_SDK_ANDROID)
   webrtc::ScopedJavaGlobalRef<jobject> android_context_;