+++ src/unity_camera_capturer_metal.mm
@@ -58,7 +58,7 @@ bool UnityCameraCapturer::MetalImpl::Init(UnityContext* context,
   return true;
 }
 
-rtc::scoped_refptr<webrtc::I420Buffer>
+webrtc::scoped_refptr<webrtc::I420Buffer>
 UnityCameraCapturer::MetalImpl::Capture() {
   auto camera_tex = (id<MTLTexture>)camera_texture_;
   auto tex = (id<MTLTexture>)frame_texture_;
@@ -102,7 +102,7 @@ UnityCameraCapturer::MetalImpl::Capture() {
                 buf.get() + width_ * 4 * (height_ - i - 1), width_ * 4);
   }
 
-  rtc::scoped_refptr<webrtc::I420Buffer> i420_buffer =
+  webrtc::scoped_refptr<webrtc::I420Buffer> i420_buffer =
       webrtc::I420Buffer::Create(width_, height_);
   libyuv::ARGBToI420(buf2.get(), width_ * 4, i420_buffer->MutableDataY(),
                      i420_buffer->StrideY(), i420_buffer->MutableDataU(),